2022/CTFZone 2022/source/bc_challenge/requirements.txt
ipython==8.4.0
iter2==1.1
returns==0.19.0
classes==0.4.1
pydantic==1.9.1
httpx==0.23.0
fastapi==0.79.0
click==8.1.3
asyncpg==0.26.0
pyjwt[crypto]==2.4.0
PyYAML==6.0